Reorganize product components.
