+++ client/src/app/modules/auth/modules/employee/components/products/new-product/new-product.component.css
