+++ client/src/app/modules/auth/modules/employee/components/products/product-detail/product-detail.component.html
